move the array test into the Test.Foundation.Array
